Remove leading spaces from fields in CSV files
In a number of places CSV files had values (or headers) with a leading
space e.g. " geo" rather than "geo". I've programmatically removed them
(and committed the script which does so).

I struggled to tell which files were generated by the scraper versus
maintained by hand - so some of this work might need to be done in the
scraper instead.